new try

+++ hashmap.c
@@ -60,8 +60,8 @@ void insertMap(HashMap * map, char * key, void * value) {
 
 void enlarge(HashMap * map) {
     enlarge_called = 1; //no borrar (testing purposes)
-    unsigned long long capacidad;
-    unsigned long long a;
+    long capacidad;
+    long a;
     Pair ** auxiliar = map -> buckets;
     map -> capacity *= 2;
     map -> buckets = (Pair**) calloc (map -> capacity , sizeof(Pair));
